+++ changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.28] - 2025-10-31
+### Added
+- No add cache to informational end-point.
+- Create a `config.py` file to centralize configurations using environment
+  variables.
+
+### Fix
+- No fix.
+
+### Changed
+- No changes.
+
+## [1.3.24] - 2025-10-31
+### Added
+- No adds.
+
+### Fix
+- Pivot column default was using order_by information.
+
+### Changed
+- No changes.
+
 ## [1.3.13] - 2025-09-19
 
 ### Added